Addressed some more comments
