+++ src/main/java/org/littleshoot/proxy/impl/BasicAuthCredentials.java
@@ -1,0 +1,20 @@
+package org.littleshoot.proxy.impl;
+
+public class BasicAuthCredentials implements Credentials {
+
+  private final String username;
+  private final String password;
+
+  BasicAuthCredentials(String username, String password) {
+    this.username = username;
+    this.password = password;
+  }
+
+  public String getUsername() {
+    return username;
+  }
+
+  public String getPassword() {
+    return password;
+  }
+}
