+++ src/main/java/org/littleshoot/proxy/impl/Credentials.java
@@ -1,20 +1,6 @@
 package org.littleshoot.proxy.impl;
 
-public class Credentials {
+public interface Credentials {
 
-  private final String username;
-  private final String password;
-
-  Credentials(String username, String password) {
-    this.username = username;
-    this.password = password;
-  }
-
-  public String getUsername() {
-    return username;
-  }
-
-  public String getPassword() {
-    return password;
-  }
+  // Marker interface
 }
